pyproject.toml
[tool.poetry]
name = "slavarobotam_task_manager"
version = "0.1.0"
description = "Light task manager"
authors = ["Mari <zefiratakuet@yandex.ru>"]
readme = "README.md"
classifiers = [
  "Private :: Do Not Upload",
]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.3"
django-bootstrap4 = "^1.1"
rollbar = "^0.14.7"
python-dotenv = "^0.10.5"
whitenoise = "^5.0.1"
gunicorn = "^20.0.4"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.4"
django-filter = "^2.2.0"


[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
